refactoring delete account

+++ java-tools/OsmAndServer/src/main/java/net/osmand/server/api/services/UserdataService.java
@@ -31,6 +31,7 @@ import javax.servlet.http.HttpServletResponse;
 import javax.transaction.Transactional;
 
 import net.osmand.server.utils.exception.OsmAndPublicApiException;
+import org.apache.commons.collections4.IterableUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -596,21 +597,31 @@ public class UserdataService {
     public ResponseEntity<String> deleteAccount(String email, PremiumUserDevicesRepository.PremiumUserDevice dev, HttpServletRequest request) throws ServletException {
         PremiumUsersRepository.PremiumUser pu = usersRepository.findByEmail(email);
         if (pu != null && pu.id == dev.userid) {
-            int numOfUsersDelete = usersRepository.deleteByEmail(email);
-            if (numOfUsersDelete != -1) {
-                int numOfUserDevicesDelete = devicesRepository.deleteByUserid(dev.userid);
-                if (numOfUserDevicesDelete != -1) {
-                    Iterable<UserFile> files = filesRepository.findAllByUserid(dev.userid);
-                    files.forEach(file -> {
-                        storageService.deleteFile(file.storage, userFolder(file), storageFileName(file));
-                        filesRepository.delete(file);
-                    });
-                    request.logout();
-                    return ResponseEntity.ok(String.format("Account has been successfully deleted (email %s)", email));
+            if (deleteAllFiles(dev)) {
+                int numOfUsersDelete = usersRepository.deleteByEmail(email);
+                if (numOfUsersDelete != -1) {
+                    int numOfUserDevicesDelete = devicesRepository.deleteByUserid(dev.userid);
+                    if (numOfUserDevicesDelete != -1) {
+                        request.logout();
+                        return ResponseEntity.ok(String.format("Account has been successfully deleted (email %s)", email));
+                    }
                 }
+                return ResponseEntity.badRequest().body(String.format("Account hasn't been deleted (%s)", email));
+            } else {
+                return ResponseEntity.badRequest().body(String.format("Unable to delete user files (%s)", email));
             }
-            return ResponseEntity.badRequest().body(String.format("Account hasn't been deleted (%s)", email));
         }
         return ResponseEntity.badRequest().body("Email doesn't match login username");
+    }
+    
+    private boolean deleteAllFiles(PremiumUserDevicesRepository.PremiumUserDevice dev) {
+        Iterable<UserFile> files = filesRepository.findAllByUserid(dev.userid);
+        files.forEach(file -> {
+            storageService.deleteFile(file.storage, userFolder(file), storageFileName(file));
+            filesRepository.delete(file);
+        });
+    
+        files = filesRepository.findAllByUserid(dev.userid);
+        return IterableUtils.size(files) == 0;
     }
 }
